Add context provider
