+++ screens/profile/ProfileScreen.tsx
@@ -12,7 +12,6 @@ import { ProfileStackParamList } from "./ProfileNavigator";
 
 import ScreenActivityIndicator from "components/ScreenActivityIndicator";
 import { View } from "components/Themed";
-import { AuthContext } from "components/AuthContext";
 import {
   EditProfileButton,
   LogoutButton,
@@ -29,6 +28,7 @@ import ProfileTabViewer from "./ProfileTabViewer";
 import Contacts from "components/profileScreen/ContactsPG";
 import AdminTab from "components/profileScreen/AdminTab";
 import MessagesTab from "components/profileScreen/MessagesTab";
+import { AuthDispatchContext } from "components/AuthContextProvider";
 
 export type ProfileScreenParams = {
   navigation: StackNavigationProp<ProfileStackParamList, "ProfileScreen">;
@@ -40,7 +40,7 @@ export default function ProfileScreen({ navigation }: ProfileScreenParams) {
   const [student, setStudent] = useState<Student | null>(null);
   const [isLoading, setLoading] = useState<boolean>(false);
   const [bookedEvents, setBookedEvents] = useState<Event[] | null>(null);
-  const authContext = useContext(AuthContext);
+  const setSignedIn = useContext(AuthDispatchContext);
   const isFocused = useIsFocused();
 
   async function getUser() {
@@ -64,8 +64,7 @@ export default function ProfileScreen({ navigation }: ProfileScreenParams) {
 
   async function logout() {
     await API.auth.logout();
-    authContext.signOut();
-    window.location.reload();
+    setSignedIn(false);
   }
 
   useFocusEffect(
